requirements.txt
aniso8601==9.0.1
blinker==1.4
click==8.1.3
email-validator==1.2.1
Flask==2.1.2
Flask-Pydantic==0.9.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.1
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.0
pandas==1.5.2
flake8==6.0.0
isort==5.10.1
